Modified some of the c/cpp syntax highlighting
By default the keyword "inline" is supposed to be highlighted as a
StorageClass (according to /usr/share/vim/vim80/syntax/c.vim). For some
reason it's marked like a constant. Therefore it was added explicitly to
c.vim

Some of the cpp keywords (inline, virtual, override, explicit, final,
export) are now also under the group StorageClass.