Use hi-default links with "!" to override defaults
Since Vim 8.2.1703 the 'hi-default' links set in
$VIMRUNTIME/syntax/<filetype>.vim are restored after ':hi clear', which
is executed when switching colorschemes with ':color {name}'. This
means, the 'hi-links' set by this plugin are lost and the user needs to
reload the buffer. The hi-links aren't loaded either when a file name is
specified on the command-line, like '$ vim test.cpp'. On the other hand,
the hi-links are loaded when the file is opened within Vim with ':edit
test.cpp'. Therefore, we need to override the 'hi-default' links with a
"!".

See also discussion in https://github.com/vim/vim/pull/6970